pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gst-python-ml"
version = "1.0.5"
description = "An ML package for GStreamer"
readme = "README.md"
requires-python = ">=3.14"
license-files = ["COPYING"]
authors = [{name = "Aaron Boxer", email = "aaron.boxer@collabora.com"}]
urls.Homepage = "https://github.com/collabora/gst-python-ml"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
    "pygobject",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "torchaudio>=2.11.0",
    "transformers>=5.6.2",
    "qwen-vl-utils[decord]>=0.0.8",
    "autoawq>=0.2.9",
    "accelerate>=1.13.0",
    "bitsandbytes>=0.49.2",
    "opencv-python>=4.13.0",
    "opencv-contrib-python>=4.13.0",
    "numpy",
    "huggingface-hub",
    "lap",
    "pycairo",
    "ultralytics",
    "confluent_kafka",
    # "flash-attn",
    "diffusers",
    "sentencepiece",
    "protobuf",
    "whisperspeech",
    "webdataset",
    "easydict",
    "pyflann-py3",
    "soundfile",
    "speechbrain",
    "pyopengl"
]

[project.optional-dependencies]
onnx = [
    "onnxruntime",
]
onnx-gpu = [
    "onnxruntime-gpu",
]
tvm = [
    "apache-tvm; python_version < '3.12'",
]
tinygrad = [
    "tinygrad",
]
mlx = [
    "mlx>=0.20.0",
    "mlx-lm",
]
executorch = [
    "executorch",
]
llamacpp = [
    "llama-cpp-python",
]
candle = [
    "candle",
]
jax-cpu = [
    "jax[cpu]",
]
jax-gpu = [
    "jax[cuda12]",
]
jax-tpu = [
    "jax[tpu]",
]
vad = [
    "pysilero",
    "pysilero-vad",
    "faster-whisper",
]

# Optional
[dependency-groups]
test = [
    "pytest>=7.0",
    # Add other test tools
]
dev = [
    {include-group = "test"},
    # Add linting, etc., e.g., "ruff"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["plugins/python"]